requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
twilio==8.10.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
requests==2.31.0
pytz==2023.3 